Added laptime minimzation
To add lap time minimzations 2 things were needed. Car variables to model the car. Velcoity simulation to control the velocity, acceleration and deceleration based on the car's variables.

Vehical:
for the vehical we store 3 main variables: friction coffient of the tires, average enginer force, and mass

Velocity:
Velocity goes through 3 phases:
1- Velocity limit:
	The velocity is limited by the amount of centriptal force and friction at each point on the track
2- Acceleration limit:
	The acceleration at each point is limited by the minimume of engine force and teraction generated by the vehicle
3- Decceleration limit:
	Decelation is limited by the traction force

for each possible path around the track the velocity is calculated for each point and the laptime is generated for the path. THe lowest laptime is picked